Fix integer truncation of small route cost values
* Actually perform a floating-point division in EdgeEvaluatorTravelTime
  to avoid unnecessary truncation of values.
* Multiply the cost value provided by edge evaluators using a constant
  factor when casting it to the integer value used in our Route.Step
  data structure to ensure that values < 1.0 are handled correctly.

Noticed-by: z29591259 @ GitHub
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>

+++ opentcs-strategies-default/src/main/java/org/opentcs/strategies/basic/routing/edgeevaluator/EdgeEvaluatorTravelTime.java
@@ -58,6 +58,6 @@ public class EdgeEvaluatorTravelTime
     else {
       maxVelocity = Math.min(vehicle.getMaxVelocity(), edge.getPath().getMaxVelocity());
     }
-    return (maxVelocity == 0) ? INFINITE_COSTS : edge.getPath().getLength() / maxVelocity;
+    return (maxVelocity == 0) ? INFINITE_COSTS : (double) edge.getPath().getLength() / maxVelocity;
   }
 }
